+++ crates/compiler/src/builtin/functions/mod.rs
@@ -10,9 +10,6 @@ use once_cell::sync::Lazy;
 
 use crate::{ast::ArgumentResult, error::SassResult, evaluate::Visitor, value::Value};
 
-#[macro_use]
-mod macros;
-
 pub mod color;
 pub mod list;
 pub mod map;