+++ crates/compiler/src/value/sass_number.rs
@@ -173,6 +173,17 @@ impl SassNumber {
     }
 
     pub fn assert_bounds(&self, name: &str, min: f64, max: f64, span: Span) -> SassResult<()> {
+        self.assert_bounds_with_unit(name, min, max, &self.unit, span)
+    }
+
+    pub fn assert_bounds_with_unit(
+        &self,
+        name: &str,
+        min: f64,
+        max: f64,
+        unit: &Unit,
+        span: Span,
+    ) -> SassResult<()> {
         if !(self.num <= Number(max) && self.num >= Number(min)) {
             return Err((
                 format!(
@@ -180,9 +191,9 @@ impl SassNumber {
                     name,
                     inspect_number(self, &Options::default(), span)?,
                     inspect_float(min, &Options::default(), span),
-                    self.unit,
+                    unit,
                     inspect_float(max, &Options::default(), span),
-                    self.unit,
+                    unit,
                 ),
                 span,
             )