+++ crates/lib/tests/math-module.rs
@@ -54,7 +54,7 @@ test!(
 test!(
     sqrt_big_positive,
     "@use 'sass:math';\na {\n  color: math.sqrt(9999999999999999999999999999999999999999999999999);\n}\n",
-    "a {\n  color: 3162277660168379038695424;\n}\n"
+    "a {\n  color: 3162277660168379000000000;\n}\n"
 );
 test!(
     sqrt_big_negative,